backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
openai==1.13.3
google-auth==2.29.0
pypdf==4.2.0
alembic==1.13.2
pytest==8.3.2
ruff==0.6.9
pre-commit==3.8.0
